==> MFD/01_mfd_price_change.ipynb <==
## Данный ноутбук рассчитывает изменение цены для каждого тикера

import pandas as pd
from tqdm import tqdm_notebook as tqdm
import sys

### Загружаем цены акций

tickers = ['mgnt', 'mtss', 'gmkn', 'sngs', 'sber', 'plzl', 'chmf', 'nvtk',
       'moex', 'afks', 'yndx', 'rsti', 'urka', 'upro', 'trnf', 'tgka',
       'hydr', 'trmk', 'rolo', 'rtkm', 'rual', 'irkt', 'sibn', 'dasb',
       'gche', 'akrn', 'arsa', 'aptk', 'amez', 'phor', 'prfn', 'kbtk',
       'pikk', 'nmtp', 'nknc', 'lsrg', 'msng', 'lnta', 'gtlc', 'yakg',
       'vtbr', 'krkn', 'rosn', 'selg', 'gaza', 'fees', 'tatn', 'svav',
       'qiwi', 'rlmn', 'lkoh', 'lsng', 'iskj', 'obuv', 'aflt', 'magn',
       'gazp', 'mtlr', 'alrs', 'rasp', 'tanl']
p = pd.DataFrame()

# читаем цену 1 тикера 
for i in tickers:
    try:
        p0 = pd.read_csv(f'/is/prices/{i.upper()}.csv', sep=',', dtype = {'<DATE>': 'str'})
        if p0.shape[1] == 1:
            p0 = pd.read_csv(f'/is/prices/{i.upper()}.csv', sep=';', dtype = {'<DATE>': 'str'})
        if p0.shape[1] == 1:
            p0 = pd.read_csv(f'/is/prices/{i.upper()}.csv', sep='|', dtype = {'<DATE>': 'str'})

        ls = []
        for c in p0.columns:
            ls.append(c.lower().replace('<', '').replace('>', ''))
        p0.columns = ls
        p = pd.concat([p, p0])
    except:
        print(f"Unexpected error with {i}", sys.exc_info()[0])
p.shape

p.info()

p.date = p.date.astype(str)
p['yea'] = p.date.apply(lambda x: x[:4])
p.head()

p1 = p[(p.yea!='2013')&(p.yea!='2014')&(p.yea!='2015')&(p.yea!='2016')&(p.yea!='2017')&(p.yea!='2018')&(p.yea!='2019')\
  &(p.yea!='2020')&(p.yea!='2021')]
p2 = p[(p.yea=='2013')|(p.yea=='2014')|(p.yea=='2015')|(p.yea=='2016')|(p.yea=='2017')|(p.yea=='2018')|(p.yea=='2019')\
  |(p.yea=='2020')|(p.yea=='2021')]

p1['d'] = p1.date.apply(lambda x: x[:2])
p1['m'] = p1.date.apply(lambda x: x[2:4])
p1['y'] = p1.date.apply(lambda x: '20' + x[4:6])
p1['date2'] = p1['y'] + p1['m'] + p1['d']
p1.head()

p1.date = p1.date2
p.columns

p1 = p1[['ticker', 'per', 'date', 'time', 'open', 'high', 'low', 'close', 'vol']]
p1.head()

p2 = p2[['ticker', 'per', 'date', 'time', 'open', 'high', 'low', 'close', 'vol']]
p2.head()

p = pd.concat([p1, p2])
p.head()

p.to_csv('/is/prices/mfd_price.csv', sep='|', index = False)

### Считаем изменение в ценах

p = pd.read_csv('/is/prices/mfd_price.csv', sep='|')
p.head()

def lag(df, price0='close', price1='close', n=1, date='date'):
    
    '''
    Функция сортирует поле price по полю date по убыванию.
    Возвращает значение доходности в день начала окна.
    
    ''' 
    
    if price0 == price1 and n > 0:
        return (df.sort_values(date, ascending=False)[price1].shift(n) / \
                df.sort_values(date, ascending=False)[price0] - 1) * 100
    
    elif price0 == price1 and n == 0:
        return (df.sort_values(date, ascending=False)[price1] / \
                df.sort_values(date, ascending=False)[price0].shift(-1) - 1) * 100


for j in [1,2,3,4,5,6,7,8,9,10, 15, 20, 25, 30, 40, 50]:
    p[f'change_close_{j}'] = None
p.head()

p = p.sort_values(['ticker', 'date'], ascending=False)
pf1 = pd.DataFrame()
pf = pd.DataFrame()
for i in tqdm(p.ticker.unique(), position=0, leave=True):
    pf = p.loc[p.ticker == i]
    for j in [1,2,3,4,5,6,7,8,9,10, 15, 20, 25, 30, 40, 50]:
        pf[f'change_close_{j}'] = lag(df=pf, price0='close', price1='close', n=j)
    pf1 = pf1.append(pf)
pf1.head()


pf1.tail()

pf1[pf1.ticker=='SBER']['change_close_1'].describe()

pf1.ticker = pf1.ticker.str.lower()
pf1.head()

# проверка
pf1.groupby('ticker')['change_close_1'].sum().sort_values(ascending=False)[:30]

pf1.describe()

pf1.to_csv('/is/prices/mfd_change_price.csv', sep='|', index = False)

pf1[(abs(pf1.change_close_1) > 5)|(abs(pf1.change_close_1) < 0.01)].shape

pf1[abs(pf1.change_close_1) < 0.05].head(30)
