==> MFD/03_mfd_label_ticker_sentiment_join.ipynb <==
## В данном ноутбуке происходит labelling акций, join сентимента и цен

import pandas as pd
import numpy as np
from tqdm.notebook import tqdm

### Сентимент

ps = pd.read_csv('/is/texts/mfd/mfd_text_lower_tic_all_flag_all_dt_e_reg_spell_stop_lem_stem_len.csv', sep='|')
ps.head()

ps.drop(columns=['per','open','high','low','close','vol','message','text','text_reg','text_reg_spell',\
                 'text_reg_spell_stop', 'day', 'month', 'len_lem_text', 'total_msg'], inplace=True)
ps.head()

print(s.shape)
s = s[(s.text_reg_spell_stop_lem != '')]
print(s.shape)
s = s[(s.text_reg_spell_stop_lem.isna() != True)]
print(s.shape)
s = s.drop_duplicates()
print(s.shape)

s['msg_init'] = 1
s.head()

### Цена

p = pd.read_csv('/is/prices/price_vol_ret_change.csv', sep='|')
p.shape

s.date = s.date.astype(float)
p.date = p.date.astype(float)
s.ticker = s.ticker.str.lower()
p.ticker = p.ticker.str.lower()

print(s.shape, p.shape)

p.rename(columns = {'init': 'price_init'}, inplace=True)
p.head()

ps = p.merge(s, how='inner', on=['date', 'ticker'])
ps.to_csv('/is/prices/tg_price_sentiment_join.csv', sep='|', index = False)
print(ps.shape)
ps.head()

print(ps.shape)
ps = ps[ps.ticker.isin(p.ticker.unique())]
print(ps.shape)

ps.columns

ps = ps[['ticker', 'date', 'open', 'high', 'low', 'close', 'vol',
       'price_init', 'volatility_5', 'volatility_20', 'volatility_120',
       'volatility_240', 'change_close_1', 'change_close_2', 'change_close_3',
       'change_close_4', 'change_close_5', 'change_close_6', 'change_close_7',
       'change_close_8', 'change_close_9', 'change_close_10',
       'change_close_15', 'change_close_20', 'change_close_25',
       'change_close_30', 'change_close_40', 'change_close_50', 'message',
       'src', 'text', 'day', 'month', 'year', 'day_of_week', 'dt', 'text_reg',
       'text_reg_spell', 'text_reg_spell_stop', 'text_reg_spell_stop_lem',
       'text_reg_spell_stop_stem', 'len_lem_text', 'total_msg', 'msg_init']]
ps.head()

ps.to_csv('/is/prices/tg_price_sentiment_join.csv', sep='|', index = False)

ps.head()

def labelling(df, price, field, n=0):

    
    '''
    Функция определяет сентимент поля price
    и указывает в новом поле field:
    1 для положительного фона,
    0 для негативного фона,
    -1 для нейтрального фона.
    Фон определяется на основе уровня изменения цены - n.
    
    '''
    df[field] = None
    
    df[field][df[price] >= (n)] = 1
    df[field][df[price] <= (-n)] = -1
    df[field][(df[price] > (-n)) & (df[price] < (n))] = 0


yields = [0.05, 0.1, 0.2, 0.3, 0.5, 0.8, 1.,  2.,  3.,  4.,  5.,  6.,  7.,  8.,  9., 10.]
lags = [1,2,3,4,5,6,7,8,9,10, 15, 20, 25, 30, 40, 50]
for lag in tqdm(lags, position=0, leave=True): 
    for income in tqdm(yields, position=0, leave=True):
        labelling(df=ps, price=f'change_close_{lag}', field=f'label_{lag}_{income}', n=income)

ps.to_csv('/is/prices/mfd_join.csv', sep='|', index = False)
ps.head()

for i in ps.columns:
    print(i)

ps.head()

ps['text_reg_spell_stop_lem'] = ps['text_reg_spell_stop_lem'].str.replace('ё', 'е')
ps['text_reg_spell_stop_stem'] = ps['text_reg_spell_stop_stem'].str.replace('ё', 'е')

ps['label_1_1.0'].dropna().astype(int).describe()

ps.head()

ps.drop(columns=['per','open','high','low','close','vol','message','text','text_reg','text_reg_spell',\
                 'text_reg_spell_stop'], inplace=True)

ps.drop(columns=['text_reg_spell_stop'], inplace=True)

ps.to_csv('/is/prices/mfd_join_lab.csv', sep='|', index = False)

ps.shape
